pyproject.toml
[tool.poetry]
name = "finalproject-christina"
version = "0.1.0"
description = "Currency trading simulation platform"
authors = ["Christina"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
prettytable = "^3.10.0"
requests = "^2.31.0"
python-dotenv = "^1.0.0"

[tool.poetry.group.dev.dependencies]
ruff = "^0.4.0"

[tool.poetry.scripts]
project = "main:main"

[tool.ruff]
line-length = 100
target-version = "py39"

[tool.ruff.lint]
select = ["E", "F", "I"]
ignore = ["E501"]

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"